# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kuben API",
    "description": "Kuben — Kubernetes-native PaaS control plane.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.1.2"
  },
  "paths": {
...
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "App detail with pods. Plain env values are included only for callers\nholding `secret-read`.",
        "operationId": "getApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetail"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apps"
        ],
        "summary": "Delete an app and everything it owns. Volumes are kept unless\n`delete_volumes=true`.",
        "operationId": "deleteApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delete_volumes",
            "in": "query",
            "description": "Also delete the app's persistent volumes. Irreversible.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "apps"
        ],
        "summary": "Update an app (image changes require `app-deploy`). Every change is a new\ndeployment run; a new tag is resolved to a digest at its registry.",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The image's registry cannot be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}/releases": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "Release history, newest first (50 revisions).",
        "operationId": "listReleases",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReleaseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}/rollback": {
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Roll back to an earlier revision: a new run of its release and\nconfiguration. The app stays on it until automatic deploys resume.",
        "operationId": "rollbackApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rollback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppDetail": {
        "type": "object",
        "required": [
          "app",
          "pods"
        ],
        "properties": {
          "app": {
            "$ref": "#/components/schemas/AppDto"
          },
          "pods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PodDto"
            }
          }
        }
      },
      "AppDto": {
        "type": "object",
        "required": [
          "name",
          "project",
          "environment",
          "namespace",
          "ready",
          "processes",
          "env",
          "domains",
          "volumes"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "description": "Environment short name."
          },
          "namespace": {
            "type": "string"
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "git_repo": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "ready": {
            "type": "boolean"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "`Available`, `Progressing`, `RolloutFailed`, `Scheduled`, `AwaitingBuild`, ..."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessDto"
            }
          },
          "env": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarDto"
            },
            "description": "Values are only included in the app detail, for callers allowed to read secrets."
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VolumeDto"
            }
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "exposure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExposureDto",
                "description": "Whether the app can be reached through the gateway, apart from\nwhether it runs (null while it has no route)."
              }
            ]
          },
          "paused": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why delivery is paused, while it is: new runs wait."
          }
        }
      },
...
      "EnvVarDto": {
        "type": "object",
        "description": "Environment variable: a plain `value` or a `secret` reference.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "LOG_LEVEL"
          },
          "value": {
            "type": [
              "string",
              "null"
            ]
          },
          "secret": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SecretRef"
              }
            ]
          }
        }
      },
...
      "ExposureDto": {
        "type": "object",
        "description": "How an app is reached through the gateway.",
        "required": [
          "hosts"
        ],
        "properties": {
          "routed": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "The gateway accepted the route and resolved its references; null\nuntil a gateway controller answered."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "hosts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HostDto"
            }
          }
        }
      },
...
      "HostDto": {
        "type": "object",
        "description": "One hostname of an app.",
        "required": [
          "host",
          "tls"
        ],
        "properties": {
          "host": {
            "type": "string"
          },
          "tls": {
            "type": "string",
            "description": "`auto` (certificate from the cluster issuer), `secret` (the app's own\ncertificate) or `none` (plain HTTP)."
          },
          "certificate_ready": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "For `auto` hosts with a certificate of their own: whether it is issued."
          },
          "certificate_message": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PodDto": {
        "type": "object",
        "required": [
          "name",
          "phase",
          "ready",
          "restarts"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "process": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "type": "string",
            "description": "`pending`, `running`, `succeeded`, `failed` or `unknown`."
          },
          "ready": {
            "type": "boolean"
          },
          "restarts": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "node": {
            "type": [
              "string",
              "null"
            ]
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Problem Details body.",
        "required": [
          "code",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `forbidden`.",
            "example": "forbidden"
          },
          "title": {
            "type": "string",
            "description": "Short human readable title."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status.",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Details, if safe to expose."
          }
        }
      },
      "ProcessDto": {
        "type": "object",
        "required": [
          "name",
          "command",
          "size",
          "min_replicas",
          "max_replicas",
          "protocol"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "string"
          },
          "min_replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "max_replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "schedule": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cron expression of scheduled processes."
          },
          "protocol": {
            "type": "string",
            "description": "`http` or `tcp`."
          }
        }
      },
...
      "ProtocolDto": {
        "type": "string",
        "enum": [
          "http",
          "tcp"
        ]
      },
...
      "ReleaseDto": {
        "type": "object",
        "required": [
          "revision",
          "reason",
          "created_at",
          "current"
        ],
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": "string",
            "description": "`create`, `deploy`, `config`, `rollback`, `promote`, `restart` or `handover`."
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "actor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Email of whoever made the change."
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "boolean",
            "description": "The newest revision (what should be running)."
          }
        }
      },
      "Rollback": {
        "type": "object",
        "required": [
          "revision"
        ],
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64",
            "description": "Revision to restore."
          }
        }
      },
...
      "SecretRef": {
        "type": "object",
        "required": [
          "name",
          "key"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        }
      },
...
      "UpdateApp": {
        "type": "object",
        "description": "Partial update; omitted fields are left unchanged, lists are replaced.",
        "properties": {
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "command": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "max_replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": [
              "string",
              "null"
            ]
          },
          "env": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/EnvVarDto"
            }
          },
          "domains": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "health_check_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Empty string removes the health check."
          },
          "schedule": {
            "type": [
              "string",
              "null"
            ],
            "description": "Empty string removes the schedule."
          },
          "time_zone": {
            "type": [
              "string",
              "null"
            ]
          },
          "protocol": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProtocolDto"
              }
            ]
          },
          "volumes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/VolumeDto"
            },
            "description": "Replaces the list. Removed volumes are unmounted, never deleted."
          },
          "fs_group": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "`0` removes the fsGroup."
          }
        }
      },
...
      "VolumeDto": {
        "type": "object",
        "description": "Persistent volume (kept when the app is deleted, unless requested).",
        "required": [
          "name",
          "mount_path"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "data"
          },
          "mount_path": {
            "type": "string",
            "example": "/data"
          },
          "size": {
            "type": "string",
            "description": "Capacity such as `5Gi`. Can grow, never shrink.",
            "example": "5Gi"
          }
        }
      },
...
    }
  },
...
}